Move `handlerfunc.NewHandler` to `httphandler.New`
