+++ httpadapter/adapter_test.go
@@ -1,4 +1,4 @@
-package handlerfunc_test
+package httpadapter_test
 
 import (
 	"fmt"
@@ -6,7 +6,7 @@ import (
 	"net/http"
 
 	"github.com/aws/aws-lambda-go/events"
-	"github.com/awslabs/aws-lambda-go-api-proxy/handlerfunc"
+	"github.com/awslabs/aws-lambda-go-api-proxy/httpadapter"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -19,17 +19,17 @@ func (h handler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 	fmt.Fprintf(w, "Go Lambda!!")
 }
 
-var _ = Describe("HandlerFuncAdapter tests", func() {
+var _ = Describe("HTTPAdapter tests", func() {
 	Context("Simple ping request", func() {
 		It("Proxies the event correctly", func() {
 			log.Println("Starting test")
 
-			handlerFunc := func(w http.ResponseWriter, req *http.Request) {
+			var httpHandler http.Handler = http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
 				w.Header().Add("unfortunately-required-header", "")
 				fmt.Fprintf(w, "Go Lambda!!")
-			}
+			})
 
-			adapter := handlerfunc.New(handlerFunc)
+			adapter := httpadapter.New(httpHandler)
 
 			req := events.APIGatewayProxyRequest{
 				Path:       "/ping",
